Move DefaultIndexURL under arduino/globals as it should be
