+++ cli/globals/globals.go
@@ -25,6 +25,4 @@ import (
 var (
 	// VersionInfo contains all info injected during build
 	VersionInfo = version.NewInfo(filepath.Base(os.Args[0]))
-	// DefaultIndexURL is the default index url
-	DefaultIndexURL = "https://downloads.arduino.cc/packages/package_index.tar.bz2"
 )
